refactor: extract stable selective rerip solver
