+++ lib/solvers/PortPointPathingSolver/tinyhypergraph/SelectiveReripTinyHyperGraphSolverWithStableInitialAssignments.ts
@@ -1,0 +1,23 @@
+import { SelectiveReripTinyHyperGraphSolver } from "tiny-hypergraph/lib/index"
+import { applyInitialAssignments } from "tiny-hypergraph/lib/initialAssignments"
+
+/**
+ * Selective rerips may move a preloaded assignment when it is the blocker.
+ * A global retry restores the serialized assignments instead of eagerly
+ * discarding every preloaded route.
+ */
+export class SelectiveReripTinyHyperGraphSolverWithStableInitialAssignments extends SelectiveReripTinyHyperGraphSolver {
+  override resetRoutingStateForRerip() {
+    super.resetRoutingStateForRerip()
+    if (!this.problem.initialAssignments?.length) return
+
+    applyInitialAssignments({
+      topology: this.topology,
+      problem: this.problem,
+      state: this.state,
+      routeSuccessCountByRouteId: this.routeSuccessCountByRouteId,
+      appendSegmentToRegionCache: (regionId, fromPortId, toPortId) =>
+        this.appendSegmentToRegionCache(regionId, fromPortId, toPortId),
+    })
+  }
+}
