+++ lib/solvers/PortPointPathingSolver/tinyhypergraph/TinyHypergraphPortPointPathingSolver.ts
@@ -17,7 +17,6 @@ import { mapLayerNameToZ } from "lib/utils/mapLayerNameToZ"
 import {
   DuplicateCongestedPortSolver,
   orderConnectionsByNetCardinality,
-  SelectiveReripTinyHyperGraphSolver,
   type DuplicateCongestedPortSolverReport,
   TinyHyperGraphSectionPipelineSolver,
   TinyHyperGraphSectionSolver,
@@ -26,7 +25,6 @@ import {
   type TinyHyperGraphSectionSolverOptions,
   type TinyHyperGraphSolverOptions,
 } from "tiny-hypergraph/lib/index"
-import { applyInitialAssignments } from "tiny-hypergraph/lib/initialAssignments"
 import type {
   ConnectionHg,
   ConnectionHgWithSimpleRouteConnection,
@@ -34,6 +32,7 @@ import type {
 } from "../hgportpointpathingsolver/types"
 import { createTinyRouteNetIndexer } from "./createTinyRouteNetIndexer"
 import { getRegionNetIdByRegionId } from "./getRegionNetIdByRegionId"
+import { SelectiveReripTinyHyperGraphSolverWithStableInitialAssignments } from "./SelectiveReripTinyHyperGraphSolverWithStableInitialAssignments"
 import {
   getSerializedPreloadedTraceStats,
   isPreloadedTraceConnectionId,
@@ -687,27 +686,6 @@ const applyMetadataPortPenalties = (loaded: LoadedTinyGraph) => {
   loaded.problem.metadataPortPenaltiesApplied = true
 
   return metadataPortPenaltyCount
-}
-
-/**
- * Selective rerips may move a preloaded assignment when it is the blocker.
- * A global retry restores the serialized assignments instead of eagerly
- * discarding every preloaded route.
- */
-class SelectiveReripTinyHyperGraphSolverWithStableInitialAssignments extends SelectiveReripTinyHyperGraphSolver {
-  override resetRoutingStateForRerip() {
-    super.resetRoutingStateForRerip()
-    if (!this.problem.initialAssignments?.length) return
-
-    applyInitialAssignments({
-      topology: this.topology,
-      problem: this.problem,
-      state: this.state,
-      routeSuccessCountByRouteId: this.routeSuccessCountByRouteId,
-      appendSegmentToRegionCache: (regionId, fromPortId, toPortId) =>
-        this.appendSegmentToRegionCache(regionId, fromPortId, toPortId),
-    })
-  }
 }
 
 class TinyHyperGraphSectionPipelineWithTerminalNetIds extends TinyHyperGraphSectionPipelineSolver {
